Fix solution for min-jumps-array
This solution in O(n).

With each step, we create a range of indexes for which we have found the
minimum steps. Using this range we calculate the max reach we can have within 1
step. If we reach the end of indices i.e. n-1, we have the steps we need
minimum. Otherwise if we have stopped moving i.e. the calculation yields the
same reach as before we stop and conclude that reaching the end is impossible.